refactor: migrate packages in other modules - remove tests module - configure compose desktop
