--- composeApp/src/commonTest/kotlin/org/example/multiplatform_demo/ComposeAppCommonTest.kt
@@ -1,12 +1,0 @@
-package org.example.multiplatform_demo
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppCommonTest {
-
-    @Test
-    fun example() {
-        assertEquals(3, 1 + 2)
-    }
-}
