+++ composeApp/src/jvmMain/kotlin/javavlsu/kb/esap/esapmobile/main.kt
@@ -1,7 +1,8 @@
-package org.example.multiplatform_demo
+package javavlsu.kb.esap.esapmobile
 
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
+import javavlsu.kb.esap.esapmobile.core.App
 
 fun main() = application {
     Window(
